+++ components/charts/BarChart.js
@@ -1,45 +1,28 @@
 import React from 'react'
-
+import PropTypes from 'prop-types'
 import {
   VictoryChart,
   VictoryBar,
   VictoryAxis,
   VictoryTooltip,
-  VictoryLabel,
-  VictoryContainer
+  VictoryLabel
 } from 'victory'
 
-import OONILogoWordmarkHoriz from '../svgs/ooni-logo-wordmark-horiz.svg'
+import OONILogoWordmarkHoriz from '../svgs/logos/OONI-HorizontalColor.svg'
 import victoryTheme from '../theme/victoryTheme'
-import { axisYStyle, baseLabelStyles } from '../theme/victoryTheme'
+import { axisYStyle } from '../theme/victoryTheme'
 import colors from '../theme/colors'
 
-class ChartTitle extends React.Component {
-  render() {
-    const { text } = this.props
-    let style = Object.assign({}, baseLabelStyles, {fontWeight: 'bold'})
-    style.fontSize = '24px'
-    return (
-      <VictoryLabel x={25} y={24} style={style} text={text} />
-    )
-  }
-}
+import ChartTitle from './ChartTitle'
 
-class ChartSubtitle extends React.Component {
-  render() {
-    const { text } = this.props
-    let style = Object.assign({}, baseLabelStyles)
-    style.fontSize = '18px'
-    return (
-      <VictoryLabel x={25} y={50} style={style} text={text} />
-    )
-  }
-}
-
-export default class extends React.Component {
+export default class BarChart extends React.Component {
 
   render () {
-    const { data, title, subtitle } = this.props
+    const {
+      data,
+      title,
+      subtitle
+    } = this.props
 
     return (
       <VictoryChart
@@ -48,7 +31,7 @@ export default class extends React.Component {
       >
 
         <ChartTitle text={title} />
-        <ChartSubtitle text={subtitle} />
+        <ChartTitle subtitle text={subtitle} />
 
         <VictoryAxis
           tickLabelComponent={
@@ -56,29 +39,35 @@ export default class extends React.Component {
           } />
         <VictoryAxis dependentAxis orientation='right' style={axisYStyle}/>
         <VictoryBar
-            labels={(d) => d.y}
-            data={data}
-            events={[{
-              target: "data",
-              eventHandlers: {
-                onMouseOver: () => null,
-                onMouseOut: () => null
-              }
-            }]}
-            labelComponent={
-              <VictoryTooltip
-                cornerRadius={0}
-                pointerLength={0}
-                height={15}
-                width={20}
-                active={true}
-                flyoutStyle={{stroke: colors.palette.base, fill: colors.palette.white}}
-              />
-            }/>
-          <g transform="translate(330, 240)">
-            <OONILogoWordmarkHoriz width={100} />
-          </g>
+          labels={(d) => d.y}
+          data={data}
+          events={[{
+            target: 'data',
+            eventHandlers: {
+              onMouseOver: () => null,
+              onMouseOut: () => null
+            }
+          }]}
+          labelComponent={
+            <VictoryTooltip
+              cornerRadius={0}
+              pointerLength={0}
+              height={15}
+              width={20}
+              active={true}
+              flyoutStyle={{stroke: colors.palette.base, fill: colors.palette.white}}
+            />
+          }/>
+        <g transform="translate(330, 240)">
+          <OONILogoWordmarkHoriz width={100} />
+        </g>
       </VictoryChart>
     )
   }
+}
+
+BarChart.propTypes = {
+  data: PropTypes.arrayOf(PropTypes.object),
+  title: PropTypes.string,
+  subtitle: PropTypes.string
 }