+++ components/components.js
@@ -1,12 +1,9 @@
-import styled from 'styled-components'
-
-import rebass from 'rebass'
-
 import atoms from './atoms'
 import molecules from './molecules'
 import organisms from './organisms'
 
 import BarChart from './charts/BarChart'
+import PieChart from './charts/PieChart'
 
 /* OONI produced atoms */
 
@@ -18,6 +15,7 @@ export const TwitterShareButton = atoms.TwitterShareButton
 export const FacebookShareButton = atoms.FacebookShareButton
 export const Heading = atoms.Heading
 export const Text = atoms.Text
+export const Select = atoms.Select
 
 /* OONI produced molecules */
 export const InputWithIconButton = molecules.InputWithIconButton
@@ -30,134 +28,77 @@ export const HeroLead = organisms.HeroLead
 export const Modal = organisms.Modal
 
 /* Here we just export rebass until we implement all of these */
-export const Absolute = rebass.Absolute
-export const Arrow = rebass.Arrow
-export const Avatar = rebass.Avatar
-export const BackgroundImage = rebass.BackgroundImage
-export const Badge = rebass.Badge
-export const Banner = rebass.Banner
-export const BlockLink = rebass.BlockLink
-export const Blockquote = rebass.Blockquote
-export const Border = rebass.Border
-export const ButtonCircle = rebass.ButtonCircle
-export const ButtonOutline = rebass.ButtonOutline
-export const ButtonTransparent = rebass.ButtonTransparent
-export const Carousel = rebass.Carousel
-export const CarouselSlide = rebass.CarouselSlide
-export const Checkbox = rebass.Checkbox
-export const Circle = rebass.Circle
-export const Close = rebass.Close
-export const Code = rebass.Code
-export const Column = rebass.Column
-export const Container = rebass.Container
-export const Divider = rebass.Divider
-export const Donut = rebass.Donut
-export const DotButton = rebass.DotButton
-export const Drawer = rebass.Drawer
-export const Embed = rebass.Embed
-export const Fixed = rebass.Fixed
-export const Group = rebass.Group
-export const Image = rebass.Image
-export const Label = rebass.Label
-export const Lead = rebass.Lead
-export const Link = rebass.Link
-export const Measure = rebass.Measure
-export const Media = rebass.Media
-export const Message = rebass.Message
-export const NavLink = rebass.NavLink
-export const Overlay = rebass.Overlay
-export const Panel = rebass.Panel
-export const PanelFooter = rebass.PanelFooter
-export const PanelHeader = rebass.PanelHeader
-export const Pre = rebass.Pre
-export const Progress = rebass.Progress
-export const Radio = rebass.Radio
-export const Relative = rebass.Relative
-export const Row = rebass.Row
-export const Samp = rebass.Samp
-export const ScrollCarousel = rebass.ScrollCarousel
-export const Select = rebass.Select
-export const Slider = rebass.Slider
-export const Small = rebass.Small
-export const Star = rebass.Star
-export const Sticky = rebass.Sticky
-export const Subhead = rebass.Subhead
-export const Switch = rebass.Switch
-export const TabItem = rebass.TabItem
-export const Tabs = rebass.Tabs
-export const Textarea = rebass.Textarea
-export const Toolbar = rebass.Toolbar
-export const Tooltip = rebass.Tooltip
-export const Truncate = rebass.Truncate
+import {
+  Absolute,
+  Arrow,
+  Avatar,
+  BackgroundImage,
+  Badge,
+  Banner,
+  BlockLink,
+  Blockquote,
+  Border,
+  ButtonCircle,
+  ButtonOutline,
+  ButtonTransparent,
+  Carousel,
+  CarouselSlide,
+  Checkbox,
+  Circle,
+  Close,
+  Code,
+  Column,
+  Container,
+  Divider,
+  Donut,
+  DotButton,
+  Drawer,
+  Embed,
+  Fixed,
+  Group,
+  Image,
+  Label,
+  Lead,
+  Link,
+  Measure,
+  Media,
+  Message,
+  NavLink,
+  Overlay,
+  Panel,
+  PanelFooter,
+  PanelHeader,
+  Pre,
+  Progress,
+  Radio,
+  Relative,
+  Row,
+  Samp,
+  ScrollCarousel,
+  Slider,
+  Small,
+  Star,
+  Sticky,
+  Subhead,
+  Switch,
+  TabItem,
+  Tabs,
+  Textarea,
+  Toolbar,
+  Tooltip,
+  Truncate
+} from 'rebass'
 
 export { default as BarChart } from './charts/BarChart'
+export { default as PieChart } from './charts/PieChart'
 
 const components = {
   Button,
-  ButtonOutline,
-  ButtonCircle,
-  ButtonTransparent,
-  Link,
-  NavLink,
-  BlockLink,
   Heading,
-  Subhead,
   Text,
-  Small,
-  Lead,
-  Pre,
-  Code,
-  Samp,
-  Blockquote,
-  Measure,
-  Truncate,
-  Label,
   Input,
   Select,
-  Textarea,
-  Checkbox,
-  Radio,
-  Slider,
-  Image,
-  Avatar,
-  BackgroundImage,
-  Container,
-  Divider,
-  Border,
-  Media,
   Card,
-  Banner,
-  Panel,
-  PanelHeader,
-  PanelFooter,
-  Progress,
-  Message,
-  Group,
-  Toolbar,
-  Badge,
-  Circle,
-  Tabs,
-  TabItem,
-  DotButton,
-  Close,
-  Relative,
-  Absolute,
-  Fixed,
-  Sticky,
-  Drawer,
-  Overlay,
-  Carousel,
-  ScrollCarousel,
-  CarouselSlide,
-  Tooltip,
-  Switch,
-  Arrow,
-  Star,
-  Embed,
-  Donut,
-  Row,
-  Column,
-  BarChart,
   RadioButton,
   RadioGroup,
   InputWithIconButton,
@@ -166,7 +107,67 @@ const components = {
   LogoOONIRun,
   TwitterShareButton,
   FacebookShareButton,
-  Modal
+  Modal,
+  BarChart,
+  PieChart,
+  Absolute,
+  Arrow,
+  Avatar,
+  BackgroundImage,
+  Badge,
+  Banner,
+  BlockLink,
+  Blockquote,
+  Border,
+  ButtonCircle,
+  ButtonOutline,
+  ButtonTransparent,
+  Carousel,
+  CarouselSlide,
+  Checkbox,
+  Circle,
+  Close,
+  Code,
+  Column,
+  Container,
+  Divider,
+  Donut,
+  DotButton,
+  Drawer,
+  Embed,
+  Fixed,
+  Group,
+  Image,
+  Label,
+  Lead,
+  Link,
+  Measure,
+  Media,
+  Message,
+  NavLink,
+  Overlay,
+  Panel,
+  PanelFooter,
+  PanelHeader,
+  Pre,
+  Progress,
+  Radio,
+  Relative,
+  Row,
+  Samp,
+  ScrollCarousel,
+  Slider,
+  Small,
+  Star,
+  Sticky,
+  Subhead,
+  Switch,
+  TabItem,
+  Tabs,
+  Textarea,
+  Toolbar,
+  Tooltip,
+  Truncate
 }
 
 export default components