+++ components/icons/NettestFacebookMessenger.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import Icon from 'react-icon-base'
+const NettestFacebookMessenger = props => (
+  <Icon viewBox="0 0 30 30" {...props}>
+    <g>
+  <defs>
+    
+  </defs>
+  <title>Facebook Messenger</title>
+  <g>
+    <path d="M15.5,17.305,12.939,14.56l-4.99,2.745,5.49-5.864,2.62,2.745,4.928-2.745Zm-1-12.6C8.953,4.7,4.455,8.9,4.455,14.061A9.115,9.115,0,0,0,8.2,21.348v3.568l3.419-1.89a10.641,10.641,0,0,0,2.882.393c5.546,0,10.044-4.193,10.044-9.358S20.045,4.7,14.5,4.7Z"/>
+  </g>
+</g>
+  </Icon>
+)
+export default NettestFacebookMessenger