+++ components/icons/Tick.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Icon from 'react-icon-base'
+const Tick = props => (
+  <Icon viewBox="0 0 30 30" {...props}>
+    <g>
+  <g>
+    <path d="M2.286,14.605l4.138-4.144c1.971,1.974,3.24,3.247,5.218,5.23L22.825,4.508l4.16,4.169L11.666,23.992Z"/>
+  </g>
+</g>
+  </Icon>
+)
+export default Tick