+++ components/theme/rebassTheme.js
@@ -3,7 +3,7 @@ import colors from './colors'
 export const rebassTheme = {
   font: '"Fira Sans", sans-serif',
   monospace: '"Source Code Pro", monospace',
-  maxWidth: 800,
+  maxWidth: 1024,
   space: [
     0,
     4,
@@ -14,10 +14,10 @@ export const rebassTheme = {
     128,
   ],
   breakpoints: [
-    32,
-    48,
-    64,
-    80
+    '32em',
+    '48em',
+    '64em',
+    '80em'
   ],
   fontSizes: [
     12,