+++ guide/content/atoms/Radio/Preview.jsx
@@ -1,6 +1,0 @@
-<RadioGroup name='animal' value='whale'>
-  <RadioButton label='Whale' value='whale' />
-  <RadioButton label='Octopus' value='octopus' />
-  <RadioButton label='Seal' value='seal' />
-</RadioGroup>
-