+++ guide/content/molecules/Card/Preview.jsx
@@ -1,3 +1,0 @@
-<Card bg='red'>
-<Heading h={1} color='white'>Lorem</Heading>
-</Card>