+++ guide/content/molecules/InputWithIconButton/Preview.jsx
@@ -1,1 +1,0 @@
-<InputWithIconButton icon={<MdDelete/>} placeholder='Enter text' />