+++ guide/content/organisms/Hero/Preview.jsx
@@ -1,4 +1,0 @@
-<Hero>
-  <HeroImg />
-  <HeroLead>Some powerful call to action</HeroLead>
-</Hero>