+++ guide/content/organisms/Modal/Preview.jsx
@@ -1,1 +1,0 @@
-<Button>Click to open modal</Button>